% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Print boilerplate stats}
\usage{
fit(measured_value, dof, pct, fitmetric = "R2", order = 6, ndecimals = 2,
  dist = "normal", table = TRUE, ...)
}
\arguments{
\item{measured_value}{a real number within the range of fitmetric}

\item{dof}{an integer}

\item{pct}{a real number between 0 and 1}

\item{fitmetric}{a character string}

\item{order}{a real number}

\item{ndecimals}{an integer}

\item{dist}{a random number distribution function}

\item{table}{a logical value}

\item{...}{any argument that functions within this routine might use}
}
\value{
data frame
}
\description{
Prints a standard set of stats for a given degree of freedom
}
\examples{
fit(6, dof=10, pct=0.95)

}

