% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTable.R
\name{fitTable}
\alias{fitTable}
\title{Print boilerplate stats}
\usage{
fitTable(dof, pctlist = c(0.9, 0.95, 0.99), ndecimals = 2,
  dist = "normal", order = 5, ...)
}
\arguments{
\item{dof}{an integer}

\item{pctlist}{a vector of real numbers between 0 and 1, minimum acceptable noise levels.}

\item{ndecimals}{an integer}

\item{dist}{a random number distribution function}

\item{order}{a real number}

\item{...}{any argument that functions within this routine might use}
}
\value{
data frame
}
\description{
Prints a standard set of stats for a given degree of freedom
}
\examples{
fitStats(6, dist='normal', sd=0.3)

}

