% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_R2df.R
\name{gen_R2df}
\alias{gen_R2df}
\title{Generate R-squared data frame of pdf and cdf values}
\usage{
gen_R2df(dof, N, bw, obs, mdl)
}
\arguments{
\item{dof}{an integer}

\item{N}{an integer}

\item{bw}{a real number}

\item{obs}{a vector of random numbers dof*N long}

\item{mdl}{a vector of random numbers dof*N long}
}
\value{
a data frame
}
\description{
Builds the pdf and cdf data frame based on random numbers generated from a specific noise distribution
}

