% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_rmsedf.R
\name{gen_userdf}
\alias{gen_userdf}
\title{Generate vector of values used to build pdf and cdf}
\usage{
gen_userdf(dof, N, bw, obs, mdl)
}
\arguments{
\item{dof}{an integer}

\item{N}{an integer}

\item{bw}{a real number describing the bin width}

\item{obs}{a vector of random numbers dof*N long}

\item{mdl}{a vector of random numbers dof*N long}
}
\value{
a vector of numbers
}
\description{
Calculates the user function output with the frequencies of each possible outcome to width bw.
}
\examples{
user <- function(m1,m2){			#user function takes two matrices, each N by dof
K <- rep(NA,nrow(m1));			#improve speed by defining the variable to store calculations
for(i in 1:nrow(m1)){
	K[i]<-cor(m1[i,],m2[i,])^2};	#For comparison, this is another method of calculating R-square
return(K)}
plotpdf(dof=3,fitmetric="user",order=5)
plotpdf(dof=3,fitmetric="R2",order=5) #for comparison

}

