% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdfs.R
\name{pcdfs}
\alias{pcdfs}
\title{Construct pdf and cdf for one of several distributions and one of several possible noise distributions}
\usage{
pcdfs(dof, order = 6, ndecimals = 2, dist = "normal", fitmetric = "R2",
  ...)
}
\arguments{
\item{dof}{an integer}

\item{order}{a real number}

\item{ndecimals}{an integer}

\item{dist}{a random number distribution function}

\item{fitmetric}{a character string naming a standard fit metric ("R2", "rmse", or "user")}

\item{...}{any argument that functions within this routine might use}
}
\value{
a data frame
}
\description{
Generates the vectors of random numbers for both observation and model and calls pcdfs
}
\examples{
pcdf(5, order=5)
pcdf(10, order=4, fitmetric="rmse")

}

