% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConstNoise.R
\name{plotConstNoise}
\alias{plotConstNoise}
\title{Plot FitValue Constant Noise}
\usage{
plotConstNoise(measured_value, dof, pct = 0.95, order = 4, plot_pctr2 = F,
  fitmetric = "R2", ...)
}
\arguments{
\item{measured_value}{a real number within the range of fitmetric}

\item{dof}{an integer}

\item{pct}{a real number between 0 and 1, minimum acceptable level of noise}

\item{order}{a real number}

\item{plot_pctr2}{a logical value indicating whether to include or not the fitEquive placed at each dof.}

\item{fitmetric}{a character string naming a standard fit metric ("R2", "rmse", or "user")}

\item{...}{any argument that functions within this routine might use}
}
\value{
ggplot object
}
\description{
Plots the Fit Equivalent
}
\examples{
plotConstNoise(0.8, 5)
plotConstNoise(0.1, 5, fitmetric="rmse")

}

