% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNoise.R
\name{plotNoise}
\alias{plotNoise}
\title{Plot Noise threshold}
\usage{
plotNoise(doflist = c(2:30), pctlist = c(0.95), order = 4,
  ndecimals = 2, fitmetric = "R2", ...)
}
\arguments{
\item{doflist}{a vector of degrees of freedom}

\item{pctlist}{a vector of percentiles (expressed as fractions)}

\item{order}{a real number greater than 0 and less than 7}

\item{ndecimals}{an integer}

\item{fitmetric}{a character string naming a standard fit metric ("R2", "rmse", or "user")}

\item{...}{any argument that functions within this routine might use}
}
\value{
ggplot object
}
\description{
Plots the Noise threshold for each degree of freedom
}
\examples{
plotNoise()

}

