% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotpdf.R
\name{plotpdf}
\alias{plotpdf}
\title{Plot PDF}
\usage{
plotpdf(dof, order = 4, dist = "normal", fitmetric = "R2", ...)
}
\arguments{
\item{dof}{an integer}

\item{order}{a real number}

\item{dist}{a random number distribution function}

\item{fitmetric}{a character string naming a standard fit metric ("R2", "rmse", or "user")}

\item{...}{any argument that functions within this routine might use}
}
\value{
ggplot object
}
\description{
Plots the Probability Density Function
}
\examples{
plotpdf(5, dist='uniform', fitmetric='RMSE')

}

