% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2.R
\name{R2}
\alias{R2}
\title{R-squared}
\usage{
R2(x, y)
}
\arguments{
\item{x}{a vector or matrix of real numbers}

\item{y}{a vector or matrix of real numbers}
}
\value{
a vector of numbers with length equal to the number of trials (N)
}
\description{
Calculates R-squared given vectors for x and y
}
\examples{
R2( seq(2:10),(3*rnorm(2:10)+4) )

}

