% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{rmse}
\usage{
rmse(obs, mdl)
}
\arguments{
\item{obs}{a vector or matrix of real numbers}

\item{mdl}{a vector or matrix of real numbers}
}
\value{
a vector of numbers with length equal to the number of trials (N)
}
\description{
Calculates rmse given vectors for y and y_pred
}
\examples{
rmse(seq(3:20), (seq(3:20)+rnorm(18)^2))

}

