% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user}
\alias{user}
\title{user}
\usage{
user(m1, m2)
}
\arguments{
\item{m1}{a vector or matrix of real numbers}

\item{m1}{a vector or matrix of real numbers}
}
\value{
a vector of numbers with length equal to the number of trials (N)
}
\description{
Calculates the user-defined functions given vectors/matrices for m1 and m2
}
\examples{
user( seq(2:10),(3*rnorm(2:10)+4) )

}

