% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utrend.R
\name{utrend}
\alias{utrend}
\title{Determine the general trend (general, overall slope) of the fitmetric function}
\usage{
utrend(fun)
}
\arguments{
\item{fun}{as function R2, rmse, or user (user-defined function)}
}
\value{
a character string ("Positive", "Negative", "Flat", "Uncertain")
}
\description{
Calculates the fitmetric function at two distant points and calculates the slope using lineal regression.
}
\examples{
utrend(R2)
utrend(user)

}

